Added scripts RayCastInteract and IInteractable
RayCastInteract is responsible for detecting interactable objects within a specified distance from the player using raycasting from a fixed camera's viewpoint. When an object that implements the IInteractable interface is detected, it is added to a list of interactable objects, allowing for potential interactions to be processed in subsequent logic.